제출/requirements.txt
pandas==2.1.3
numpy==1.26.0
koreanize-matplotlib==0.1.1
matplotlib==3.8.2
selenium==4.19.0
statsmodels==0.14.3
yfinance==0.2.38
scikit-learn==1.3.2
seaborn==0.13.1
faiss-cpu==1.8.0
sentence-transformers==2.6.1
openai == 0.28.0
yake==0.4.8
openpyxl==3.1.4
requests==2.31.0
beautifulsoup4==4.12.2
pip == 24.2
SQLAlchemy == 2.0.36
pyodbc == 5.2.0
PySocks == 1.7.1
tf_keras == 2.18.0